Move getRandomChars() to @console/shared so it can be used globally
