+++ frontend/packages/dev-console/src/utils/shared-submit-utils.ts
@@ -145,10 +145,3 @@ export const createRoute = (
 
   return route;
 };
-
-export const getRandomChars = (digit = 6): string => {
-  return Math.random()
-    .toString(36)
-    .replace(/[^a-z0-9]+/g, '')
-    .substr(1, digit);
-};
